refactor: fold native provider proxy into cli api
